No need to remove an entry log, if it reported an error.

+++ routes/workflow.js
@@ -210,9 +210,8 @@ function status(workflow) {
 
     let seconds_begin = Math.floor((now - workflow.task.begin)); // The difference between now and the last call from the agent.
     let seconds_end = Math.floor((now - workflow.task.end)); // The difference between now and the last call from the agent.
-    let running_time = seconds_end - seconds_begin;
 
-    console.log("Workflow status " + workflow.fileset + ':' + workflow.task.queue + ':' + workflow.task.status + ':' + workflow.task.retry + ':' + seconds_begin + ':' + seconds_end + ':' + running_time);
+    console.log("Workflow status " + workflow.fileset + ':' + workflow.task.queue + ':' + workflow.task.status + ':' + workflow.task.retry + ':' + seconds_begin + ':' + seconds_end + ':' + workflow.task.retryTime);
     switch (workflow.task.status) {
         case 100:
         case 150:
@@ -262,9 +261,11 @@ function status(workflow) {
             save(workflow);
             break;
         case 499:
-            if (workflow.task.retry && running_time > workflow.task.retry) { // For hour no response yet
+            if (workflow.task.retryTime < 0) { // For hour no response yet
                 console.log("Retry task: " + workflow.task.queue);
                 amq(workflow);
+            } else {
+                console.log("Failed task. Time before retry " + workflow.task.retryTime);
             }
             break;
         case 500:
@@ -308,7 +309,7 @@ router.put('/heartbeat', function (req, res) {
 
     if ( run_heartbeat++ !== 1 ) {
         res.status(200);
-        res.end(JSON.stringify({status: 200, message: 'Still running heartbeat...' + run_heartbeat}));
+        res.end(JSON.stringify({status: 200, message: 'Still running heartbeat... ' + run_heartbeat}));
         return;
     }
 
@@ -316,21 +317,20 @@ router.put('/heartbeat', function (req, res) {
     let statuschecked = []; // something to report
     let deleted = []; // something to delete
 
-    const query = {};
-    Workflow.find(query).stream()
+    Workflow.find().stream()
         .on('error', function (err) {
             res.status(500);
             res.end(JSON.stringify({status: 500, message: err}));
         })
         .on('data', function (workflow) {
 
+            statuschecked.push(workflow.fileset);
             let _workflow = Workflow(workflow);
             if (fs.existsSync(workflow.fileset)) {
                 console.log("Heartbeat status check for workflow " + workflow.fileset);
-                statuschecked.push(workflow.fileset);
                 status(_workflow);
             } else {
-                if ( workflow.complete === false || workflow.delete_on_success) {
+                if ( workflow.complete === false && workflow.task.status !== 499 || workflow.delete_on_success) {
                     console.log("Fileset no longer on filesystem. Removing workflow " + workflow.fileset);
                     deleted.push(workflow.fileset);
                     _workflow.delete();
